Replacing all 'unsigned long' instances with 'size_t' and 'long' with 'sys::SSize_T'.  This fixed at least one bug that occurred on Windows in NITFReadControl::interleaved() with AOI's > 4 GB.  Also fixed a bug in test_dump_images when it was run with doing a single read - it was allocating and writing out more than was needed.  #306.
git-svn-id: https://wush.net/svn/mdaus-oss/trunk@485 9d772353-a6c9-4ee9-b2cc-df47a7649cd5

+++ modules/c++/six.sicd/include/six/sicd/ComplexData.h
@@ -165,7 +165,7 @@ public:
      *  Maps to: /SICD/ImageData/NumRows,
      *  /SIDD/Measurement/PixelFootprint/Row
      */
-    virtual unsigned long getNumRows() const
+    virtual size_t getNumRows() const
     {
         return imageData->numRows;
     }
@@ -174,7 +174,7 @@ public:
      *  the full image rows, and the scene center rows
      *
      */
-    virtual void setNumRows(unsigned long numRows)
+    virtual void setNumRows(size_t numRows)
     {
         imageData->numRows = numRows;
         imageData->fullImage.row = numRows;
@@ -185,7 +185,7 @@ public:
      *  Maps to: /SICD/ImageData/NumCols,
      *  /SIDD/Measurement/PixelFootprint/Col
      */
-    virtual unsigned long getNumCols() const
+    virtual size_t getNumCols() const
     {
         return imageData->numCols;
     }
@@ -194,7 +194,7 @@ public:
      *  the full image columns, and the scene center column
      *
      */
-    virtual void setNumCols(unsigned long numCols)
+    virtual void setNumCols(size_t numCols)
     {
         imageData->numCols = numCols;
         imageData->fullImage.col = numCols;
