+++ modules/c++/six/include/six/NITFSegmentInfo.h
@@ -35,16 +35,16 @@ namespace six
 struct NITFSegmentInfo
 {
     //! First row in the image segment in real space
-    unsigned long firstRow;
+    size_t firstRow;
 
     //! Row offset in the CCS (ILOC R)
     //  When you are attached to another segment, ILOC is with respect to that
     //  segment.  Per the spec, we will always attach to the previous segment,
     //  so this will simply be the number of rows in that previous segment.
-    unsigned long rowOffset;
+    size_t rowOffset;
 
     //! Number of rows in this segment
-    unsigned long numRows;
+    size_t numRows;
 
     //! The image segment corner points
     LatLonCorners corners;
